Adds customization function
